Added examples for boost asio library
Added some boost asio library example to study later.  Will use
asio to communicate between the module and the module engine I think
since this will allow us to run each module on a different thread and
not have to open and close the program over and over.  The question
then is whether this can be extended to other languages.