Added beginning Module framework
Each Module class will take care of managing 1 module executable and
format file.  The ModuleEngine class will then keep a list of Module
classes.  The Module class will encapsulate the reading of the Module
format files (however we decide to format those) as well as the
location of the executable file.